chore: split context and provider for fast reload
