+++ src/CommandContext.ts
@@ -1,0 +1,9 @@
+import { createContext } from "react"
+
+export const CommandContext = createContext<{
+    exec: (command: string, data: object) => void
+}>({
+    exec: () => () => {
+        throw new Error("Missing CommandProvider")
+    },
+})
